# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reminders": {
      "post": {
        "operationId": "RemindersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReminderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Required agent assignment is missing"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Agents cannot reassign reminders"
          },
          "404": {
            "description": "Agent not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a reminder",
        "tags": [
          "reminders"
        ]
      },
      "get": {
        "operationId": "RemindersController_all",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReminderResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "The authenticated user has no agent profile"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List reminders visible to the current user",
        "tags": [
          "reminders"
        ]
      }
    },
    "/reminders/{id}": {
      "patch": {
        "operationId": "RemindersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReminderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Closed reminders cannot be edited"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Agents can only update their own reminders"
          },
          "404": {
            "description": "Reminder or agent not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a pending reminder",
        "tags": [
          "reminders"
        ]
      }
    },
    "/reminders/{id}/complete": {
      "patch": {
        "operationId": "RemindersController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteReminderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Reminder is not pending or the follow-up is invalid"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Agents can only complete their own reminders"
          },
          "404": {
            "description": "Reminder not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Complete a pending reminder",
        "tags": [
          "reminders"
        ]
      }
    },
    "/reminders/{id}/dismiss": {
      "patch": {
        "operationId": "RemindersController_dismiss",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Only pending reminders can be dismissed"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Agents can only dismiss their own reminders"
          },
          "404": {
            "description": "Reminder not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dismiss a pending reminder",
        "tags": [
          "reminders"
        ]
      }
    },
...
    "/volunteers/{volunteerId}/availability": {
      "post": {
        "operationId": "VolunteerAvailabilityController_create",
        "parameters": [
          {
            "name": "volunteerId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVolunteerAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolunteerAvailabilityResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "End time must be after start time"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Volunteers can only manage their own availability"
          },
          "404": {
            "description": "Volunteer not found"
          },
          "409": {
            "description": "Volunteer is inactive or the slot overlaps an existing slot"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a volunteer availability slot",
        "tags": [
          "volunteer-availability"
        ]
      },
      "get": {
        "operationId": "VolunteerAvailabilityController_findAll",
        "parameters": [
          {
            "name": "volunteerId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VolunteerAvailabilityResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Volunteers can only view their own availability"
          },
          "404": {
            "description": "Volunteer not found"
          },
          "409": {
            "description": "Volunteer is inactive"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List a volunteer availability slots",
        "tags": [
          "volunteer-availability"
        ]
      }
    },
    "/volunteers/{volunteerId}/availability/{id}": {
      "delete": {
        "operationId": "VolunteerAvailabilityController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "volunteerId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Slot removed"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Volunteers can only manage their own availability"
          },
          "404": {
            "description": "Volunteer or availability slot not found"
          },
          "409": {
            "description": "Volunteer is inactive or availability is reserved"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a volunteer availability slot",
        "tags": [
          "volunteer-availability"
        ]
      }
    },
...
    "/volunteers/{id}": {
      "get": {
        "operationId": "VolunteersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolunteerResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "404": {
            "description": "Volunteer not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a volunteer",
        "tags": [
          "volunteers"
        ]
      },
      "patch": {
        "operationId": "VolunteersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVolunteerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolunteerResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Administrator role required"
          },
          "404": {
            "description": "Volunteer not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a volunteer",
        "tags": [
          "volunteers"
        ]
      }
    },
...
  },
  "info": {
    "title": "FPC Backend API",
    "description": "FPC backend API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateReminderDto": {
        "type": "object",
        "properties": {
          "subjectPatientId": {
            "type": "string",
            "format": "uuid"
          },
          "dueAt": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "assignedAgentId": {
            "type": "string",
            "format": "uuid"
          },
          "createdFromFollowUpId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "subjectPatientId",
          "dueAt",
          "description"
        ]
      },
      "ReminderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subjectPatientId": {
            "type": "string",
            "format": "uuid"
          },
          "createdFromFollowUpId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "assignedAgentId": {
            "type": "string",
            "format": "uuid"
          },
          "dueAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING",
              "DONE",
              "DISMISSED"
            ],
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "resultingFollowUpId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subjectPatientId",
          "createdFromFollowUpId",
          "assignedAgentId",
          "dueAt",
          "description",
          "status",
          "completedAt",
          "resultingFollowUpId",
          "createdAt"
        ]
      },
...
      "UpdateReminderDto": {
        "type": "object",
        "properties": {
          "assignedAgentId": {
            "type": "string",
            "format": "uuid"
          },
          "dueAt": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CompleteReminderDto": {
        "type": "object",
        "properties": {
          "resultingFollowUpId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CreateVolunteerAvailabilityDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          }
        },
        "required": [
          "date",
          "startTime",
          "endTime"
        ]
      },
      "VolunteerAvailabilityResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "volunteerId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "example": "09:00:00"
          },
          "endTime": {
            "type": "string",
            "example": "10:00:00"
          },
          "status": {
            "enum": [
              "AVAILABLE",
              "RESERVED"
            ],
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "volunteerId",
          "date",
          "startTime",
          "endTime",
          "status",
          "createdAt"
        ]
      },
...
      "VolunteerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "firstName",
          "lastName",
          "specialty",
          "email",
          "phone",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateVolunteerDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "maxLength": 255
          },
          "lastName": {
            "type": "string",
            "maxLength": 255
          },
          "specialty": {
            "type": "string",
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "maxLength": 50
          }
        }
      }
    }
  }
}